refactor(api): move usecase and test to Team domain
