+++ api/src/team/application/certification-center-membership/certification-center-membership.admin.controller.js
@@ -1,4 +1,3 @@
-import { usecases as libUsecases } from '../../../../lib/domain/usecases/index.js';
 import { BadRequestError } from '../../../shared/application/http-errors.js';
 import * as certificationCenterMembershipSerializer from '../../../shared/infrastructure/serializers/jsonapi/certification-center-membership.serializer.js';
 import { requestResponseUtils } from '../../../shared/infrastructure/utils/request-response-utils.js';
@@ -64,7 +63,7 @@ const findCertificationCenterMembershipsByUser = async function (
   dependencies = { certificationCenterMembershipSerializer },
 ) {
   const userId = request.params.id;
-  const certificationCenterMemberships = await libUsecases.findCertificationCenterMembershipsByUser({
+  const certificationCenterMemberships = await usecases.findCertificationCenterMembershipsByUser({
     userId,
   });
   return h.response(
@@ -76,7 +75,7 @@ const certificationCenterMembershipAdminController = {
   updateRole,
   disableFromPixAdmin,
   createCertificationCenterMembershipByEmail,
-  findCertificationCenterMembershipsByUser
+  findCertificationCenterMembershipsByUser,
 };
 
 export { certificationCenterMembershipAdminController };
