+++ api/src/team/domain/usecases/find-certification-center-memberships-by-user.js
